+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 9.6.1
 - wasm-crypto 6.1.5
 - api, api-contract, types 8.9.2-9
 - hw-ledger, keyring, util, util-crypto 9.5.2-15